% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychTestPerformance.R
\docType{package}
\name{psychTestPerformance-package}
\alias{psychTestPerformance}
\alias{psychTestPerformance-package}
\title{psychTestPerformance: Collection of the LongGold PsychTestR performance tests and questionnaires}
\description{
Implementation of LongGold test battery consisting of a selection of performance tests and questionnaires based on the psychTestR library. Demos of the whole set of performance tests and questionnaires can be accessed at https://shiny.gold-msi.org/longgold_demo .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmhoeger/psychTestPerformance}
  \item Report bugs at \url{https://github.com/fmhoeger/psychTestPerformance/issues}
}

}
\keyword{internal}
