% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{psychTestPerformance_conflicts}
\alias{psychTestPerformance_conflicts}
\title{This function lists all the conflicts between packages in psychTestPerformance
and other packages that you have loaded.}
\usage{
psychTestPerformance_conflicts()
}
\description{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
psychTestPerformance_conflicts()
}
