% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{psychTestPerformance_deps}
\alias{psychTestPerformance_deps}
\title{List all psychTestPerformance dependencies}
\usage{
psychTestPerformance_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
psychTestPerformance packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all psychTestPerformance dependencies
}
