% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psychTestPerformance_packages}
\alias{psychTestPerformance_packages}
\title{List all packages in the psychTestPerformance}
\usage{
psychTestPerformance_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include psychTestPerformance in the list?}
}
\description{
List all packages in the psychTestPerformance
}
\examples{
psychTestPerformance_packages()
}
