% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psyperform.R
\docType{package}
\name{psyperform-package}
\alias{psyperform}
\alias{psyperform-package}
\title{psyperform: LongGold psychTestR performance tests}
\description{
Implementation of a LongGold test battery consisting of performance tests based on the psychTestR library. Demos of the whole set of performance tests, listening tests, and questionnaires can be accessed at https://shiny.gold-msi.org/longgold_demo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmhoeger/psyperform}
  \item Report bugs at \url{https://github.com/fmhoeger/psyperform/issues}
}

}
\keyword{internal}
