% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{psyperform_deps}
\alias{psyperform_deps}
\title{List all psyperform dependencies}
\usage{
psyperform_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
psyperform packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all psyperform dependencies
}
