% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psyperform_packages}
\alias{psyperform_packages}
\title{List all packages in the psyperform}
\usage{
psyperform_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include psyperform in the list?}
}
\description{
List all packages in the psyperform
}
\examples{
psyperform_packages()
}
