% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{psyperform_update}
\alias{psyperform_update}
\title{Update psyperform packages}
\usage{
psyperform_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
psyperform packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all psyperform packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
psyperform_update()
}
}
